+++ common/src/main/scala/explore/components/ui/ExploreStyles.scala
@@ -396,7 +396,6 @@ object ExploreStyles {
   val AladinFullScreenButton       = Css("explore-aladin-fullscreen-button")
   val AladinSettingsMenu           = Css("explore-aladin-settings-menu")
   val BlendedSVG                   = Css("blended-svg")
-  val AgsCalculationMessage        = Css("guide-star-calculation")
 
   val ExploreToast     = Css("explore-toast")
   val ExploreToastGrid = Css("explore-toast-grid")
