+++ project/Settings.scala
@@ -40,7 +40,6 @@ object Settings {
     val refinedAlgebraVersion = "0.1.0"
     val scalaJsReact          = "2.1.1"
     val webAppUtil            = "2.0.0-RC2"
-    val boopickle             = "1.4.0"
   }
 
   object Libraries {
