+++ workers/src/main/scala/workers/CatalogCache.scala
@@ -145,10 +145,7 @@ trait CatalogCache extends CatalogIDB with AsyncToIO {
                   _.collect { case Right(s) =>
                     GuideStarCandidate.siderealTarget.get(s)
                   }.map(gsc =>
-                    // We keep locally the data already pm corrected for the viz time
-                    // If it changes over a month we'll request the data again and recalculate
-                    // This way we avoid recalculating pm for example if only pos angle or
-                    // conditions change
+                    // Do PM correction
                     gsc.at(request.vizTime)
                   )
                 )
